+++ packages/compiler/test/css_parser/BUILD.bazel
@@ -1,0 +1,30 @@
+load("//tools:defaults.bzl", "jasmine_node_test", "ts_library", "ts_web_test_suite")
+
+ts_library(
+    name = "css_parser_lib",
+    testonly = True,
+    srcs = glob(["**/*.ts"]),
+    deps = [
+        "//packages:types",
+        "//packages/compiler/testing",
+        "//packages/core/testing",
+        "//packages/platform-browser",
+        "//packages/platform-browser/testing",
+    ],
+)
+
+jasmine_node_test(
+    name = "css_parser",
+    bootstrap = ["angular/tools/testing/init_node_spec.js"],
+    deps = [
+        ":css_parser_lib",
+        "//tools/testing:node",
+    ],
+)
+
+ts_web_test_suite(
+    name = "css_parser_web",
+    deps = [
+        ":css_parser_lib",
+    ],
+)
