+++ packages/compiler/test/expression_parser/BUILD.bazel
@@ -1,0 +1,30 @@
+load("//tools:defaults.bzl", "jasmine_node_test", "ts_library", "ts_web_test_suite")
+
+ts_library(
+    name = "expression_parser_lib",
+    testonly = True,
+    srcs = glob(["**/*.ts"]),
+    deps = [
+        "//packages:types",
+        "//packages/compiler",
+        "//packages/compiler/test/expression_parser/utils",
+        "//packages/compiler/testing",
+        "//packages/platform-browser/testing",
+    ],
+)
+
+jasmine_node_test(
+    name = "expression_parser",
+    bootstrap = ["angular/tools/testing/init_node_spec.js"],
+    deps = [
+        ":expression_parser_lib",
+        "//tools/testing:node",
+    ],
+)
+
+ts_web_test_suite(
+    name = "expression_parser_web",
+    deps = [
+        ":expression_parser_lib",
+    ],
+)
