+++ packages/compiler/test/i18n/translation_bundle_spec.ts
@@ -12,7 +12,7 @@ import * as i18n from '../../src/i18n/i18n_ast';
 import {TranslationBundle} from '../../src/i18n/translation_bundle';
 import * as html from '../../src/ml_parser/ast';
 import {ParseLocation, ParseSourceFile, ParseSourceSpan} from '../../src/parse_util';
-import {serializeNodes} from '../ml_parser/ast_serializer_spec';
+import {serializeNodes} from '../ml_parser/util/util';
 
 import {_extractMessages} from './i18n_parser_spec';
