+++ packages/compiler/test/ml_parser/BUILD.bazel
@@ -1,0 +1,28 @@
+load("//tools:defaults.bzl", "jasmine_node_test", "ts_library", "ts_web_test_suite")
+
+ts_library(
+    name = "ml_parser_lib",
+    testonly = True,
+    srcs = glob(["**/*.ts"]),
+    deps = [
+        "//packages:types",
+        "//packages/compiler",
+        "//packages/compiler/test/ml_parser/util",
+    ],
+)
+
+jasmine_node_test(
+    name = "ml_parser",
+    bootstrap = ["angular/tools/testing/init_node_spec.js"],
+    deps = [
+        ":ml_parser_lib",
+        "//tools/testing:node",
+    ],
+)
+
+ts_web_test_suite(
+    name = "ml_parser_web",
+    deps = [
+        ":ml_parser_lib",
+    ],
+)
